chore: move environment variables from `Environment.java` to `build.gradle` using BuildConfig fields chore: package cleanup chore: update build.gradle
