+++ app/src/main/java/com/cottacush/android/baseProject/BackButtonFragment.java
@@ -1,4 +1,4 @@
-package com.example.testing.myapplication.base;
+package com.cottacush.android.baseProject;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
