+++ app/src/main/java/com/cottacush/android/baseProject/BaseFragment.java
@@ -1,4 +1,4 @@
-package com.example.testing.myapplication.base;
+package com.cottacush.android.baseProject;
 
 import android.annotation.SuppressLint;
 import android.app.DatePickerDialog;
