+++ app/src/main/java/com/cottacush/android/baseProject/BasePresenter.java
@@ -1,4 +1,4 @@
-package com.example.testing.myapplication.base;
+package com.cottacush.android.baseProject;
 
 /**
  * Created by Jubril on 9/19/17.
