+++ app/src/main/java/com/cottacush/android/baseProject/MainActivity.java
@@ -2,6 +2,7 @@ package com.cottacush.android.baseProject;
 
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
+import android.util.Log;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -9,6 +10,8 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+
+        Log.d("TESTING Build config", BuildConfig.ENVIRONMENT);
     }
 
 }
