--- app/src/staging/java/com/cottacush/android/baseProject/Environment.java
@@ -1,8 +1,0 @@
-package com.cottacush.android.baseProject;
-
-/**
- * Created by becca on 11/30/17.
- */
-
-public class Environment {
-}
